Integrated Components
Integrated available components to give the web app a seem less flow between pages. Set the foundation for authentication using context. Reconfigured tailwind.config.js cause it was altered and this broke our dependencies. Made the Sidebar and header as well as the collapse functionality it has an routing to other pages and keeping them in the position of outlet.